Replace goal_found exception with nullptr test.
OK, I sinfully used an exception to jump out of a deeply-nested recursive function,
which was amusing but wrong.  It is now replaced by manually unwinding the recursion
and returning nullptr if no goal is found or a goal node if one is found.  This
makes the interface a bit C-like, which is not ideal but it's OK for now.